Create sensor module
